FEAT :: 커스텀 스크롤, 파일 구조 변경
