+++ components/layout/section/project/ProjectItem.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function ProjectItem() {
+  return <div className="h-48">ProjectItem</div>;
+}
+
+export default ProjectItem;
