+++ pages/project/index.tsx
@@ -1,5 +1,5 @@
 import Header from "@/components/layout/header/Header";
-import ProjectSection from "@/components/layout/section/project";
+import ProjectSection from "@/components/layout/section/project/Project";
 import Sidebar from "@/components/layout/sidebar/Sidebar";
 import React from "react";
