+++ styles/globals.css
@@ -8,7 +8,19 @@
   }
   body {
     font-size: 1.5rem;
-    @apply text-text bg-lightBackground dark:bg-black;
+    @apply text-text bg-lightBackground dark:bg-black overflow-hidden;
+  }
+
+  *::-webkit-scrollbar {
+    @apply w-[2rem];
+  }
+
+  *::-webkit-scrollbar-thumb {
+    @apply rounded-[1rem] dark:bg-textDarkGray bg-sidebarLightText mr-10 w-[0.4rem] bg-clip-padding border-transparent border-8 border-solid;
+  }
+
+  *::-webkit-scrollbar-track {
+    @apply bg-lightBack dark:bg-darkGray;
   }
 }
 
@@ -17,6 +29,6 @@
     @apply active:scale-90 dark:hover:bg-darkHover hover:bg-lightHover duration-250 rounded-xl dark:text-textLightGray cursor-pointer inline-flex items-center justify-center;
   }
   .main-section {
-    @apply rounded-tl-2xl absolute top-[5.4rem] left-[25rem] bg-lightBack dark:bg-darkGray flex items-center justify-center flex-col w-[calc(100vw-25rem)] h-[calc(100vh-5.4rem)] font-bold text-4xl dark:text-white;
+    @apply rounded-tl-2xl absolute top-[5.4rem] left-[25rem] bg-lightBack dark:bg-darkGray w-[calc(100vw-25rem)] min-h-[calc(100vh-5.4rem)] dark:text-white;
   }
 }
